Clarify comments around minijail_pre{enter|exec}().
I was investigating Minijail behaviour with static binaries and
it took me a while to figure out what was happening where.

Document preenter/preexec functions better and move them closer
to the flags they track. This way if we add a new flag in the future
we'll also track it in minijail_pre{enter|exec}().

BUG=None
TEST=unit, security_{Minijail0|Minijail_seccomp} on leon.

Change-Id: I67e1e233b0fa0df2dcd97ad397187a7dc791a0c3
Reviewed-on: https://chromium-review.googlesource.com/194200
Reviewed-by: Kees Cook <keescook@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>