PM: CopyVariable can determine whether the source value was set.
CopyVariable is now capable of checking an (optional) flag to see
whether the referenced value is set, instead of blindly copying it. This
preserves the existing interface, which defaults to an "always set"
state. The general interface also accepts a error message that is
returned to the caller of GetValue() in the case where a requested value
was not set.

BUG=None
TEST=Unit tests.

Change-Id: I983ea7b8c3985853cecce7a7b44f208b8b968e21
Reviewed-on: https://chromium-review.googlesource.com/195594
Tested-by: Gilad Arnold <garnold@chromium.org>
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Gilad Arnold <garnold@chromium.org>