shill: DHCPConfig: Reorder callback and lease expiration timer
The lease expiration timer should be started before calling the
"UpdateProperties" method, since the latter may end up stopping
the DHCPConfig instance.  Otherwise, we may inadevertently start
an expiration timer with no running DHCP process.

The concept of having an expiration timer is attractive, so that
even with a static IP configuration (the plausible reason that
the config would be stopped) it might be attractive to
periodically restart the lease acquisition process.  However it
probably makes sense to use a different timer for this purpose.

BUG=chromium:364735
TEST=Unit tests, repro of bug steps

Change-Id: I252dbeb29fc9daf6f2522ebed8c3886df2b50bb7
Reviewed-on: https://chromium-review.googlesource.com/195560
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>