homedirs.cc: Ignore unused authorized keys
At present, only the exact key type may be added for authorized updates.
This means it is impossible to provide keys in advance of use.  For
instance, upgrading from a HMAC-only solution to an encrypt+HMAC
method.  This change enables pre-loading additional keys and updates the
unittest to exercise the path that was failing -- signing key first.

BUG=chromium:364883
TEST=builds and new unittest passes (with both key types)

Change-Id: I0aa0453d3fce46e261c94b10a7427d04e6e7640a
Reviewed-on: https://chromium-review.googlesource.com/196361
Reviewed-by: Will Drewry <wad@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>
Commit-Queue: Will Drewry <wad@chromium.org>