shill: Persist user updates to AutoConnect for non-ephemeral services.
When a user sets the AutoConnect property, we take it as a clue that we should
persist the user preference in the profile associated with the service. Even if
the property update does not change the property itself, we should flush the
property to the profile, to ensure we persist the (unchanged) value.

BUG=chromium:364633
TEST=(1) Run shill unit-tests.
     (2) network_3GModemControl.pseudomodem passes consistently.

Change-Id: I11fbb6e85ce9fc4fa80f809c0b21b8e8d0cd0495
Reviewed-on: https://chromium-review.googlesource.com/196390
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>