PolicyManager: EvaluationContext caches null values.
When a variable fails during an evaluation, we don't want to retry it
on that evaluation. That would imply that the same policy could get
more than one value for that variable (one null and one not null)
which can make bugs harder to find.

Const variables that fail, should fail during the whole program, and
reflect that a given subsystem isn't present on the build.

BUG=None
TEST=unittest modified.

Change-Id: Icb69623318d31883e0a857ce0d1fb900ca68290f
Reviewed-on: https://chromium-review.googlesource.com/196704
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>