shill: IPConfig: Don't persist data
IPConfig objects in shill are only useful for reporting state
over the RPC API.  They are read-only, and no state is loaded
or used.  At the start of shill development an effort was made
to have IPConfigs save their "Type" property as well as for
the Device object to save an "IPConfigs" property that
referred to them.  This may have been in order to maintain
some level of file-compatibility with flimflam, but even this
is no longer necessary.

This CL removes the Load and Save methods from IPConfigs, and
rips away everything that depended on it.

BUG=None
TEST=Unit tests, manual: start shill

Change-Id: I5b5aa562c015d1f9b069dd28dd662022bb1e7240
Reviewed-on: https://chromium-review.googlesource.com/197160
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>