Unit tests: Use Md5sum to decide whether to discard MMAPs
Previously, we used the array index of the test files, and discarded
MMAPs for even-indexed files.

This became a problem when we wanted to insert new files into the list
of test files as the following files would get shifted by one. To make
this more robust, use Md5sum prefixes instead of array indexes to
decide. This should yield a 50/50 mix of discard vs no-discard and is
immune to adding or removing files.

BUG=chromium:367372
TEST=unit tests pass

Change-Id: I009be2b65a7faf1a78cd6976dceb6455e50b0ef9
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/198575