shill: OpenVPNDriver: Allow option to ignore default route
In OpenVPN, it is up to the client to decide whether or not to
accept a default route from the OpenVPN server.  If any routes
are specified by the server, the "route_vpn_gateway" option
may be provided in order to specify the default router to be
used in pushed routes.  This does not necessarily mean that
the server also wishes that the client adds a default route
through this gateway.  There is no pushed configuration from
the server that specifies what this behavior should be.

By default, ChromeOS creates a default route to the gateway.
This CL adds a configuration option to VPN services which allows
users or administrators wishing to have a "split tunel"
configuration to omit the default route option.  Supplying this
option does not override explicit routes pushed by the server, so
administrators can still effectively force a single route through
their chosen gateway.

CQ-DEPEND=CL:198900
BUG=chromium:370460
TEST=Unit tests

Change-Id: I8a3efe717b0d70742065aa9b89671ecd21f3f613
Reviewed-on: https://chromium-review.googlesource.com/198902
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>