UploadCrashes: run in the main mount namespace
When debugd launches, it places itself into a new mount namespace at
boot time.  That means that any tools it invokes won't have access to
all the user mounts.  Normally this is a good thing, but this makes
sending crashes a lot less useful to the end user as they often want
to trigger user crashes in addition to system crashes.

For now, wrap the crash_sender call with nsenter so that we run in
the same mount namespace as init (which should be the same namespace
that user mounts are made in).  We probably want this logic to get
merged into minijail in some way.

BUG=chromium:376639
TEST=generated a chrome crash; invoked new crash_sender via debugd; saw user crash get uploaded
TEST=`cbuildbot x86-generic-full` works

Change-Id: I9a3e2673ed285daab16dcc301b069f7fccd8769a
Reviewed-on: https://chromium-review.googlesource.com/201374
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>