key.proto: add KeyProviderData and renamed CROS_LEGACY
Add KeyProviderData to KeyData such that the provider of a key can
store the metadata that it needs at a future time to regenerate the
key, such as the salt or a pre-hashing algorithm, and so on.

Additionally, KeyType KEY_TYPE_PASSWORD_CROS_LEGACY was renamed to
KEY_TYPE_PASSWORD to indicate that a secret is used (as opposed to a
asymmetric one).  CROS_LEGACY as a name was a misstep because it
leaked client information (hashing approach) into the Key type
rather than making it a client-specific attribute, e.g., via
KeyProviderData or a KeyProviderType attribute.  cryptohome only
needs to be aware that the key type means a secret will be presented
in order to unwrap the data it protects.

(As there are no references in Chromium or cryptohome to the
 original type, the renaming should have no perceivable impact.)

TEST=builds
BUG=chromium:367847

Change-Id: Iaba2c665add9a30b839da69d354e4d1643c38fd8
Reviewed-on: https://chromium-review.googlesource.com/201730
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Commit-Queue: Will Drewry <wad@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>